sms_email/callmebot_whatsapp_example/requirements.txt
Flask==2.3.3
requests==2.32.4
python-dotenv==1.0.0
